requirements.txt
Django==3.2.9
djangorestframework==3.12.4
Pillow==8.4.0
qrcode==7.3.1